Update HedString further, use in more places.  Update validation to only work on long tags.  Misc other cleanup and fixes
