+++ hedtools/examples/run_hed_conversion_examples.py
@@ -4,38 +4,40 @@ This example is specifically converting tags between short and long form.
 
 Classes Demonstrated:
 HedFileInput - Used to open/modify/save a spreadsheet
-TagFormat - Used to convert between short and long tags
+HedSchema - Used to convert hed strings between short and long forms
 """
 import os
 
 from hed.util.hed_file_input import HedFileInput
-from hed.tools.tag_format import TagFormat
+from hed.schema.hed_schema_file import load_schema
 
 local_hed_file_no_dupe = 'data/HED8.0.0-alpha.1.xml'
 
 
-def long_to_short_file(input_file):
-    tag_formatter = TagFormat(local_hed_file_no_dupe)
-    converted_file, errors = tag_formatter.convert_file_to_short_tags(input_file)
-    converted_file.save(include_formatting=True, add_suffix="_test_long_to_short")
+def long_to_short_file(input_file, hed_schema, error_handler=None):
+    error_list = input_file.convert_to_short(hed_schema, error_handler)
+    input_file.save(include_formatting=True, add_suffix="_test_long_to_short")
+    return input_file, error_list
 
 
-def short_to_long_file(input_file):
-    tag_formatter = TagFormat(local_hed_file_no_dupe)
-    converted_file, errors = tag_formatter.convert_file_to_long_tags(input_file)
-    converted_file.save(include_formatting=True, add_suffix="_test_short_to_long")
+def short_to_long_file(input_file, hed_schema, error_handler=None):
+    error_list = input_file.convert_to_long(hed_schema, error_handler)
+    input_file.save(include_formatting=True, add_suffix="_test_short_to_long")
+    return input_file, error_list
+
 
 if __name__ == '__main__':
     example_data_path = 'data'
     hed3_tags_single_sheet = os.path.join(example_data_path, 'hed3_tags_single_sheet.xlsx')
 
-    prefixed_needed_tag_columns = {2: 'Event/Label/', 3: 'Event/Description/'}
-    input_file = HedFileInput(hed3_tags_single_sheet, tag_columns=[4],
+    loaded_schema = load_schema(local_hed_file_no_dupe)
+    prefixed_needed_tag_columns = {2: 'Attribute/Informational/Label/', 3: 'Attribute/Informational/Description/'}
+    loaded_file = HedFileInput(hed3_tags_single_sheet, tag_columns=[4],
                               column_prefix_dictionary=prefixed_needed_tag_columns,
                               worksheet_name='LKT Events')
-    long_to_short_file(input_file)
+    long_to_short_file(loaded_file, loaded_schema)
 
-    input_file = HedFileInput(hed3_tags_single_sheet, tag_columns=[4],
+    loaded_file = HedFileInput(hed3_tags_single_sheet, tag_columns=[4],
                               column_prefix_dictionary=prefixed_needed_tag_columns,
                               worksheet_name='LKT Events')
-    short_to_long_file(input_file)
+    short_to_long_file(loaded_file, loaded_schema)
