+++ hedtools/examples/run_wiki2xml.py
@@ -11,7 +11,7 @@ import hed
 from hed import schema
 
 if __name__ == '__main__':
-    hed_wiki_url = 'https://raw.githubusercontent.com/hed-standard/hed-specification/master/HED-generation2-schema-V2.0.0.mediawiki'
+    hed_wiki_url = 'https://raw.githubusercontent.com/hed-standard/hed-specification/master/HED-generation2-schema-V7.2.0.mediawiki'
     local_hed_file = None
     xml_location, errors = schema.convert_schema_to_format(hed_wiki_url, local_hed_file=local_hed_file)
     hed.get_printable_issue_string(validation_issues=errors, title="Errors in HED-generation3-schema.mediawiki")
@@ -19,7 +19,7 @@ if __name__ == '__main__':
         move(xml_location, "outputGen2.xml")
 
     try:
-        hed_wiki_url = 'https://raw.githubusercontent.com/hed-standard/hed-specification/master/HED-generation3-schema-V2.0.0.mediawiki'
+        hed_wiki_url = 'https://raw.githubusercontent.com/hed-standard/hed-specification/master/HED-generation3-schema-V8.0.0-alpha.2.mediawiki'
         local_hed_file = None
         xml_location, errors = schema.convert_schema_to_format(hed_wiki_url, local_hed_file=local_hed_file)
         hed.get_printable_issue_string(validation_issues=errors, title="Errors in HED-generation3-schema.mediawiki")
