+++ hedtools/tests/tools/test_tag_format.py
@@ -2,8 +2,9 @@ import unittest
 import os
 
 from hed.util.error_types import SchemaErrors
-from hed.tools.tag_format import TagFormat
 from hed.util import error_reporter
+from hed.util.hed_string import HedString
+from hed.schema.hed_schema_file import load_schema
 
 
 class TestTagFormat(unittest.TestCase):
@@ -12,34 +13,43 @@ class TestTagFormat(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         hed_xml = os.path.join(os.path.dirname(os.path.abspath(__file__)), cls.schema_file)
-        cls.tag_compare = TagFormat(hed_xml)
+        cls.hed_schema = load_schema(hed_xml)
         cls.error_handler = error_reporter.ErrorHandler()
 
 
 class TestConvertTag(TestTagFormat):
-    def validator_base(self, test_function, test_strings, expected_results, expected_errors):
+    def validator_base(self, test_strings, expected_results, expected_errors, convert_to_short=True):
         for test_key in test_strings:
-            test_result, test_errors = test_function(test_strings[test_key])
+            test_string_obj = HedString(test_strings[test_key])
+            if convert_to_short:
+                test_errors = test_string_obj.convert_to_short(self.hed_schema)
+            else:
+                test_errors = test_string_obj.convert_to_long(self.hed_schema)
             expected_error = expected_errors[test_key]
             expected_result = expected_results[test_key]
-            self.assertEqual(test_result, expected_result, test_strings[test_key])
+            self.assertEqual(str(test_string_obj), expected_result, test_strings[test_key])
             self.assertCountEqual(test_errors, expected_error, test_strings[test_key])
 
 
 class TestConvertString(TestTagFormat):
-    def validator_base(self, test_function, test_strings, expected_results, expected_errors):
+    def validator_base(self, test_strings, expected_results, expected_errors, convert_to_short=True):
         for test_key in test_strings:
-            test_result, test_errors = test_function(test_strings[test_key])
+            test_string_obj = HedString(test_strings[test_key])
+            if convert_to_short:
+                test_errors = test_string_obj.convert_to_short(self.hed_schema)
+            else:
+                test_errors = test_string_obj.convert_to_long(self.hed_schema)
             expected_error = expected_errors[test_key]
             expected_result = expected_results[test_key]
-            self.assertEqual(test_result, expected_result, test_strings[test_key])
+            self.assertEqual(str(test_string_obj), expected_result, test_strings[test_key])
             self.assertCountEqual(test_errors, expected_error, test_strings[test_key])
 
 
 class TestConvertToLongTag(TestConvertTag):
     def validator(self, test_strings, expected_results, expected_errors):
-        super(TestConvertToLongTag, self).validator_base(self.tag_compare._convert_to_long_tag, test_strings,
-                                                         expected_results, expected_errors)
+        super(TestConvertToLongTag, self).validator_base(test_strings,
+                                                         expected_results, expected_errors,
+                                                         convert_to_short=False)
 
     def test_tag(self):
         test_strings = {
@@ -89,14 +99,11 @@ class TestConvertToLongTag(TestConvertTag):
             'trailingSpace': 'Environmental-sound/Unique Value ',
         }
         expected_results = {
-            'leadingSpace': ' Environmental-sound/Unique Value',
-            'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value ',
+            'leadingSpace': 'Item/Sound/Environmental-sound/Unique Value',
+            'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value',
         }
         expected_errors = {
-            'leadingSpace':
-                self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND, test_strings['leadingSpace'],
-                                                            0, 20)
-            ,
+            'leadingSpace': [],
             'trailingSpace': [],
         }
         self.validator(test_strings, expected_results, expected_errors)
@@ -243,8 +250,8 @@ class TestConvertToLongTag(TestConvertTag):
 
 class TestConvertToShortTag(TestConvertTag):
     def validator(self, test_strings, expected_results, expected_errors):
-        super(TestConvertToShortTag, self).validator_base(self.tag_compare._convert_to_short_tag, test_strings,
-                                                          expected_results, expected_errors)
+        super(TestConvertToShortTag, self).validator_base(test_strings,
+                                                          expected_results, expected_errors, convert_to_short=True)
 
     def test_tag(self):
         test_strings = {
@@ -306,13 +313,11 @@ class TestConvertToShortTag(TestConvertTag):
             ,
             'multiLevel':
                 self.error_handler.format_schema_error(SchemaErrors.INVALID_PARENT_NODE,
-                                                            test_strings['multiLevel'], 37, 50,
-                                                 'Event/Sensory-event')
+                                                            test_strings['multiLevel'], 31, 36,'Event')
             ,
             'mixed':
                 self.error_handler.format_schema_error(SchemaErrors.INVALID_PARENT_NODE,
-                                                            test_strings['mixed'], 31, 50,
-                                                 'Item/Sound/Environmental-sound')
+                                                            test_strings['mixed'], 11, 16, 'Event')
             ,
         }
         self.validator(test_strings, expected_results, expected_errors)
@@ -323,14 +328,11 @@ class TestConvertToShortTag(TestConvertTag):
             'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value ',
         }
         expected_results = {
-            'leadingSpace': ' Item/Sound/Environmental-sound/Unique Value',
-            'trailingSpace': 'Environmental-sound/Unique Value ',
+            'leadingSpace': 'Environmental-sound/Unique Value',
+            'trailingSpace': 'Environmental-sound/Unique Value',
         }
         expected_errors = {
-            'leadingSpace':
-                self.error_handler.format_schema_error(SchemaErrors.INVALID_PARENT_NODE, test_strings['leadingSpace'],
-                                                            12, 31, 'Item/Sound/Environmental-sound')
-            ,
+            'leadingSpace': [],
             'trailingSpace': [],
         }
         self.validator(test_strings, expected_results, expected_errors)
@@ -379,9 +381,9 @@ class TestConvertToShortTag(TestConvertTag):
                                                             test_strings['singleLevelAlreadyShort'],
                                                             19, 28, 'Item/Object/Geometric'),
             'twoLevels': self.error_handler.format_schema_error(SchemaErrors.INVALID_PARENT_NODE,
-                                                                      test_strings['twoLevels'], 35, 40, 'Event'),
+                                                                      test_strings['twoLevels'], 25, 34, 'Item/Object/Geometric'),
             'duplicate': self.error_handler.format_schema_error(SchemaErrors.INVALID_PARENT_NODE,
-                                                                      test_strings['duplicate'], 34, 43, 'Item/Object/Geometric')
+                                                                      test_strings['duplicate'], 22, 26, 'Item')
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -402,28 +404,16 @@ class TestConvertToShortTag(TestConvertTag):
         }
         expected_errors = {
             'invalidParentWithExistingGrandchild': self.error_handler.format_schema_error(
-                SchemaErrors.INVALID_PARENT_NODE,
-                test_strings[
-                                                                                         'invalidParentWithExistingGrandchild'],
-                32, 41,
-                                                                                     'Item/Object/Geometric'),
+                SchemaErrors.NO_VALID_TAG_FOUND, test_strings['invalidParentWithExistingGrandchild'], 0, 12),
             'invalidChildWithExistingGrandchild': self.error_handler.format_schema_error(
-                SchemaErrors.INVALID_PARENT_NODE,
-                test_strings[
-                                                                                        'invalidChildWithExistingGrandchild'],
-                19, 28,
-                                                                                    'Item/Object/Geometric'),
+                SchemaErrors.INVALID_PARENT_NODE, test_strings['invalidChildWithExistingGrandchild'], 19, 28,
+                "Item/Object/Geometric"),
             'invalidParentWithExistingChild': self.error_handler.format_schema_error(
-                SchemaErrors.INVALID_PARENT_NODE,
-                test_strings[
-                                                                                    'invalidParentWithExistingChild'],
-                13, 22,
-                                                                                'Item/Object/Geometric'),
-            'invalidSingle': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                          test_strings['invalidSingle'], 0, 12),
-            'invalidWithExtension':
-                self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                            test_strings['invalidWithExtension'], 0, 12),
+                SchemaErrors.NO_VALID_TAG_FOUND, test_strings['invalidParentWithExistingChild'], 0, 12),
+            'invalidSingle': self.error_handler.format_schema_error(
+                SchemaErrors.NO_VALID_TAG_FOUND, test_strings['invalidSingle'], 0, 12),
+            'invalidWithExtension': self.error_handler.format_schema_error(
+                SchemaErrors.NO_VALID_TAG_FOUND, test_strings['invalidWithExtension'], 0, 12),
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -495,8 +485,8 @@ class TestConvertToShortTag(TestConvertTag):
 
 class TestConvertHedStringToShort(TestConvertString):
     def validator(self, test_strings, expected_results, expected_errors):
-        super(TestConvertHedStringToShort, self).validator_base(self.tag_compare.convert_hed_string_to_short,
-                                                                test_strings, expected_results, expected_errors)
+        super(TestConvertHedStringToShort, self).validator_base(test_strings, expected_results, expected_errors,
+                                                                convert_to_short=True)
 
     def test_empty_strings(self):
         test_strings = {
@@ -506,7 +496,8 @@ class TestConvertHedStringToShort(TestConvertString):
             'emptyString': '',
         }
         expected_errors = {
-            'emptyString': self.error_handler.format_schema_error(SchemaErrors.EMPTY_TAG_FOUND, '')
+            #'emptyString': self.error_handler.format_schema_error(SchemaErrors.EMPTY_TAG_FOUND, '')
+            'emptyString': []
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -514,24 +505,24 @@ class TestConvertHedStringToShort(TestConvertString):
         test_strings = {
             'singleLevel': 'Event',
             'multiLevel': 'Event/Sensory-event',
-            'twoSingle': 'Event, Attribute',
+            'twoSingle': 'Event,Attribute',
             'oneExtension': 'Event/Extension',
-            'threeMulti': 'Event/Sensory-event, Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5',
-            'simpleGroup': '(Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5)',
-            'groupAndTag': '(Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5), Item/Object/Man-made/Vehicle/Car',
-            'nestedGroup': '((Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5), Item/Object/Man-made/Vehicle/Car, Attribute/Environmental/Indoors)',
-            'nestedGroup2': '(Item/Object/Man-made/Vehicle/Car, Attribute/Environmental/Indoors, (Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5))'
+            'threeMulti': 'Event/Sensory-event,Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5',
+            'simpleGroup': '(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5)',
+            'groupAndTag': '(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5),Item/Object/Man-made/Vehicle/Car',
+            'nestedGroup': '((Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5),Item/Object/Man-made/Vehicle/Car,Attribute/Environmental/Indoors)',
+            'nestedGroup2': '(Item/Object/Man-made/Vehicle/Car,Attribute/Environmental/Indoors,(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5))'
         }
         expected_results = {
             'singleLevel': 'Event',
             'multiLevel': 'Sensory-event',
-            'twoSingle': 'Event, Attribute',
+            'twoSingle': 'Event,Attribute',
             'oneExtension': 'Event/Extension',
-            'threeMulti': 'Sensory-event, Train, RGB-red/0.5',
-            'simpleGroup': '(Train, RGB-red/0.5)',
-            'groupAndTag': '(Train, RGB-red/0.5), Car',
-            'nestedGroup': '((Train, RGB-red/0.5), Car, Indoors)',
-            'nestedGroup2': '(Car, Indoors, (Train, RGB-red/0.5))'
+            'threeMulti': 'Sensory-event,Train,RGB-red/0.5',
+            'simpleGroup': '(Train,RGB-red/0.5)',
+            'groupAndTag': '(Train,RGB-red/0.5),Car',
+            'nestedGroup': '((Train,RGB-red/0.5),Car,Indoors)',
+            'nestedGroup2': '(Car,Indoors,(Train,RGB-red/0.5))'
         }
         expected_errors = {
             'singleLevel': [],
@@ -552,16 +543,16 @@ class TestConvertHedStringToShort(TestConvertString):
         test_strings = {
             'single': single,
             'double': double,
-            'both': single + ', ' + double,
-            'singleWithTwoValid': 'Attribute, ' + single + ', Event',
-            'doubleWithValid': double + ', Item/Object/Man-made/Vehicle/Car/Minivan',
+            'both': single + ',' + double,
+            'singleWithTwoValid': 'Attribute,' + single + ',Event',
+            'doubleWithValid': double + ',Item/Object/Man-made/Vehicle/Car/Minivan',
         }
         expected_results = {
             'single': single,
             'double': double,
-            'both': single + ', ' + double,
-            'singleWithTwoValid': 'Attribute, ' + single + ', Event',
-            'doubleWithValid': double + ', Car/Minivan',
+            'both': single + ',' + double,
+            'singleWithTwoValid': 'Attribute,' + single + ',Event',
+            'doubleWithValid': double + ',Car/Minivan',
         }
         expected_errors = {
             'single': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
@@ -569,13 +560,13 @@ class TestConvertHedStringToShort(TestConvertString):
             'double': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
                                                                    test_strings['double'], 0, 12),
             'both': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                 test_strings['both'], 0, 12)
+                                                           single, 0, 12)
                      + self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                 test_strings['both'], 14, 26),
+                                                              double, 0, 12),
             'singleWithTwoValid': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                               test_strings['singleWithTwoValid'], 11, 23),
+                                                                         single, 0, 12),
             'doubleWithValid': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                            test_strings['doubleWithValid'], 0, 12),
+                                                                      double, 0, 12),
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -584,17 +575,17 @@ class TestConvertHedStringToShort(TestConvertString):
             'leadingSpace': ' Item/Sound/Environmental-sound/Unique Value',
             'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value ',
             'bothSpace': ' Item/Sound/Environmental-sound/Unique Value ',
-            'leadingSpaceTwo': ' Item/Sound/Environmental-sound/Unique Value, Event',
-            'trailingSpaceTwo': 'Event, Item/Sound/Environmental-sound/Unique Value ',
-            'bothSpaceTwo': ' Event, Item/Sound/Environmental-sound/Unique Value ',
+            'leadingSpaceTwo': ' Item/Sound/Environmental-sound/Unique Value,Event',
+            'trailingSpaceTwo': 'Event,Item/Sound/Environmental-sound/Unique Value ',
+            'bothSpaceTwo': ' Event,Item/Sound/Environmental-sound/Unique Value ',
         }
         expected_results = {
-            'leadingSpace': ' Environmental-sound/Unique Value',
-            'trailingSpace': 'Environmental-sound/Unique Value ',
-            'bothSpace': ' Environmental-sound/Unique Value ',
-            'leadingSpaceTwo': ' Environmental-sound/Unique Value, Event',
-            'trailingSpaceTwo': 'Event, Environmental-sound/Unique Value ',
-            'bothSpaceTwo': ' Event, Environmental-sound/Unique Value ',
+            'leadingSpace': 'Environmental-sound/Unique Value',
+            'trailingSpace': 'Environmental-sound/Unique Value',
+            'bothSpace': 'Environmental-sound/Unique Value',
+            'leadingSpaceTwo': 'Environmental-sound/Unique Value,Event',
+            'trailingSpaceTwo': 'Event,Environmental-sound/Unique Value',
+            'bothSpaceTwo': 'Event,Environmental-sound/Unique Value',
         }
         expected_errors = {
             'leadingSpace': [],
@@ -703,8 +694,8 @@ class TestConvertHedStringToShort(TestConvertString):
 
 class TestConvertHedStringToLong(TestConvertString):
     def validator(self, test_strings, expected_results, expected_errors):
-        super(TestConvertHedStringToLong, self).validator_base(self.tag_compare.convert_hed_string_to_long,
-                                                               test_strings, expected_results, expected_errors)
+        super(TestConvertHedStringToLong, self).validator_base(test_strings, expected_results, expected_errors,
+                                                               convert_to_short=False)
 
     def test_empty_strings(self):
         test_strings = {
@@ -714,7 +705,8 @@ class TestConvertHedStringToLong(TestConvertString):
             'emptyString': '',
         }
         expected_errors = {
-            'emptyString': self.error_handler.format_schema_error(SchemaErrors.EMPTY_TAG_FOUND, '')
+            #'emptyString': self.error_handler.format_schema_error(SchemaErrors.EMPTY_TAG_FOUND, '')
+            'emptyString': []
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -722,24 +714,24 @@ class TestConvertHedStringToLong(TestConvertString):
         test_strings = {
             'singleLevel': 'Event',
             'multiLevel': 'Sensory-event',
-            'twoSingle': 'Event, Attribute',
+            'twoSingle': 'Event,Attribute',
             'oneExtension': 'Event/Extension',
-            'threeMulti': 'Sensory-event, Train, RGB-red/0.5',
-            'simpleGroup': '(Train, RGB-red/0.5)',
-            'groupAndTag': '(Train, RGB-red/0.5), Car',
-            'nestedGroup': '((Train, RGB-red/0.5), Car, Indoors)',
-            'nestedGroup2': '(Car, Indoors, (Train, RGB-red/0.5))'
+            'threeMulti': 'Sensory-event,Train,RGB-red/0.5',
+            'simpleGroup': '(Train,RGB-red/0.5)',
+            'groupAndTag': '(Train,RGB-red/0.5),Car',
+            'nestedGroup': '((Train,RGB-red/0.5),Car,Indoors)',
+            'nestedGroup2': '(Car,Indoors,(Train,RGB-red/0.5))'
         }
         expected_results = {
             'singleLevel': 'Event',
             'multiLevel': 'Event/Sensory-event',
-            'twoSingle': 'Event, Attribute',
+            'twoSingle': 'Event,Attribute',
             'oneExtension': 'Event/Extension',
-            'threeMulti': 'Event/Sensory-event, Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5',
-            'simpleGroup': '(Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5)',
-            'groupAndTag': '(Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5), Item/Object/Man-made/Vehicle/Car',
-            'nestedGroup': '((Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5), Item/Object/Man-made/Vehicle/Car, Attribute/Environmental/Indoors)',
-            'nestedGroup2': '(Item/Object/Man-made/Vehicle/Car, Attribute/Environmental/Indoors, (Item/Object/Man-made/Vehicle/Train, Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5))'
+            'threeMulti': 'Event/Sensory-event,Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5',
+            'simpleGroup': '(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5)',
+            'groupAndTag': '(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5),Item/Object/Man-made/Vehicle/Car',
+            'nestedGroup': '((Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5),Item/Object/Man-made/Vehicle/Car,Attribute/Environmental/Indoors)',
+            'nestedGroup2': '(Item/Object/Man-made/Vehicle/Car,Attribute/Environmental/Indoors,(Item/Object/Man-made/Vehicle/Train,Attribute/Sensory/Visual/Color/RGB-color/RGB-red/0.5))'
         }
         expected_errors = {
             'singleLevel': [],
@@ -760,16 +752,16 @@ class TestConvertHedStringToLong(TestConvertString):
         test_strings = {
             'single': single,
             'double': double,
-            'both': single + ', ' + double,
-            'singleWithTwoValid': 'Attribute, ' + single + ', Event',
-            'doubleWithValid': double + ', Car/Minivan',
+            'both': single + ',' + double,
+            'singleWithTwoValid': 'Attribute,' + single + ',Event',
+            'doubleWithValid': double + ',Car/Minivan',
         }
         expected_results = {
             'single': single,
             'double': double,
-            'both': single + ', ' + double,
-            'singleWithTwoValid': 'Attribute, ' + single + ', Event',
-            'doubleWithValid': double + ', Item/Object/Man-made/Vehicle/Car/Minivan',
+            'both': single + ',' + double,
+            'singleWithTwoValid': 'Attribute,' + single + ',Event',
+            'doubleWithValid': double + ',Item/Object/Man-made/Vehicle/Car/Minivan',
         }
         expected_errors = {
             'single': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
@@ -777,13 +769,13 @@ class TestConvertHedStringToLong(TestConvertString):
             'double': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
                                                                    test_strings['double'], 0, 12),
             'both': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                 test_strings['both'], 0, 12)
+                                                                 single, 0, 12)
                      + self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                 test_strings['both'], 14, 26),
+                                                                 double, 0, 12),
             'singleWithTwoValid': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                               test_strings['singleWithTwoValid'], 11, 23),
+                                                                               single, 0, 12),
             'doubleWithValid': self.error_handler.format_schema_error(SchemaErrors.NO_VALID_TAG_FOUND,
-                                                                            test_strings['doubleWithValid'], 0, 12),
+                                                                            double, 0, 12),
         }
         self.validator(test_strings, expected_results, expected_errors)
 
@@ -792,17 +784,17 @@ class TestConvertHedStringToLong(TestConvertString):
             'leadingSpace': ' Environmental-sound/Unique Value',
             'trailingSpace': 'Environmental-sound/Unique Value ',
             'bothSpace': ' Environmental-sound/Unique Value ',
-            'leadingSpaceTwo': ' Environmental-sound/Unique Value, Event',
-            'trailingSpaceTwo': 'Event, Environmental-sound/Unique Value ',
-            'bothSpaceTwo': ' Event, Environmental-sound/Unique Value ',
+            'leadingSpaceTwo': ' Environmental-sound/Unique Value,Event',
+            'trailingSpaceTwo': 'Event,Environmental-sound/Unique Value ',
+            'bothSpaceTwo': ' Event,Environmental-sound/Unique Value ',
         }
         expected_results = {
-            'leadingSpace': ' Item/Sound/Environmental-sound/Unique Value',
-            'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value ',
-            'bothSpace': ' Item/Sound/Environmental-sound/Unique Value ',
-            'leadingSpaceTwo': ' Item/Sound/Environmental-sound/Unique Value, Event',
-            'trailingSpaceTwo': 'Event, Item/Sound/Environmental-sound/Unique Value ',
-            'bothSpaceTwo': ' Event, Item/Sound/Environmental-sound/Unique Value ',
+            'leadingSpace': 'Item/Sound/Environmental-sound/Unique Value',
+            'trailingSpace': 'Item/Sound/Environmental-sound/Unique Value',
+            'bothSpace': 'Item/Sound/Environmental-sound/Unique Value',
+            'leadingSpaceTwo': 'Item/Sound/Environmental-sound/Unique Value,Event',
+            'trailingSpaceTwo': 'Event,Item/Sound/Environmental-sound/Unique Value',
+            'bothSpaceTwo': 'Event,Item/Sound/Environmental-sound/Unique Value',
         }
         expected_errors = {
             'leadingSpace': [],
