+++ hedtools/tests/validator/test_hed_validator.py
@@ -24,9 +24,9 @@ class Test(unittest.TestCase):
         cls.comma_delimited_list_with_double_quotes = ['a', 'b', 'c', "d,e,f"]
         cls.comma_delimiter = ','
         cls.category_key = 'Category'
-        cls.hed_string_with_multiple_unique_tags = 'event/label/this is a label,event/label/this is another label'
-        cls.hed_string_with_invalid_tags = 'this/is/not/a/valid/tag1,this/is/not/a/valid/tag2'
-        cls.hed_string_with_no_required_tags = 'no/required/tags1,no/required/tags'
+        cls.hed_string_with_multiple_unique_tags = HedString('event/label/this is a label,event/label/this is another label')
+        cls.hed_string_with_invalid_tags = HedString('this/is/not/a/valid/tag1,this/is/not/a/valid/tag2')
+        cls.hed_string_with_no_required_tags = HedString('no/required/tags1,no/required/tags')
         cls.attribute_onset_tag = 'Attribute/Onset'
         cls.category_partipant_and_stimulus_tags = 'Event/Category/Participant response,Event/Category/Stimulus'
         cls.category_tags = 'Participant response, Stimulus'
@@ -51,23 +51,19 @@ class Test(unittest.TestCase):
         self.assertTrue(display_filename in validation_issues[0][ErrorContext.FILE_NAME])
 
     def test__validate_individual_tags_in_hed_string(self):
-        hed_string_delimiter = HedString(self.hed_string_with_invalid_tags)
-        validation_issues = self.generic_hed_input_reader._validate_individual_tags_in_hed_string(hed_string_delimiter)
+        validation_issues = self.generic_hed_input_reader._validate_individual_tags_in_hed_string(self.hed_string_with_invalid_tags)
         self.assertIsInstance(validation_issues, list)
         self.assertTrue(validation_issues)
 
     def test__validate_top_levels_in_hed_string(self):
-        hed_string_delimiter = HedString(self.hed_string_with_no_required_tags)
-        validation_issues = self.generic_hed_input_reader._validate_tags_in_hed_string(hed_string_delimiter)
+        validation_issues = self.generic_hed_input_reader._validate_tags_in_hed_string(self.hed_string_with_no_required_tags)
         self.assertIsInstance(validation_issues, list)
         self.assertFalse(validation_issues)
 
     def test__validate_tag_levels_in_hed_string(self):
-        hed_string_delimiter = HedString(self.hed_string_with_multiple_unique_tags)
-        validation_issues = self.generic_hed_input_reader._validate_tag_levels_in_hed_string(hed_string_delimiter)
+        validation_issues = self.generic_hed_input_reader._validate_tag_levels_in_hed_string(self.hed_string_with_multiple_unique_tags)
         self.assertIsInstance(validation_issues, list)
         self.assertTrue(validation_issues)
-
 
     def test__append_validation_issues_if_found(self):
         row_number = random.randint(0, 100)
